+++ src/App.scss
@@ -1,4 +1,4 @@
-@import './src/Utilities.scss';
+@import './src/utilities.scss';
 @import '~@redhat-cloud-services/frontend-components-notifications/index.css';
 @import '~@redhat-cloud-services/frontend-components-sources/index.css';
