--- src/Utilities.scss
@@ -1,8 +1,0 @@
-// Sass Imports
-
-@import 'Utilities/_variables.scss';
-@import 'Utilities/_colors.scss';
-@import 'Utilities/_helpers.scss';
-@import 'Utilities/_mixins.scss';
-
-@import '~@redhat-cloud-services/frontend-components/index.css';
