+++ src/entries-dev.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import App from './App';
-import { getDevStore } from './Utilities/getDevStore';
+import { getDevStore } from './utilities/getDevStore';
 
 export const DevEntry = () => (<Provider store={getDevStore()}>
     <App />
