+++ src/test/components/redirectNotAdmin/RedirectNotAdmin.spec.js
@@ -3,7 +3,7 @@ import configureStore from 'redux-mock-store';
 import { mount } from 'enzyme';
 import { act } from 'react-dom/test-utils';
 
-import { componentWrapperIntl } from '../../../Utilities/testsHelpers';
+import { componentWrapperIntl } from '../../../utilities/testsHelpers';
 import * as actions from '../../../redux/sources/actions';
 import RedirectNotAdmin from '../../../components/RedirectNotAdmin/RedirectNotAdmin';
 import { routes, replaceRouteId } from '../../../Routes';
