+++ src/test/components/sourceEditForm/SourceEditModal.spec.js
@@ -7,7 +7,7 @@ import thunk from 'redux-thunk';
 import { act } from 'react-dom/test-utils';
 import { Spinner } from '@patternfly/react-core/dist/js/components/Spinner';
 
-import { componentWrapperIntl } from '../../../Utilities/testsHelpers';
+import { componentWrapperIntl } from '../../../utilities/testsHelpers';
 import SourceEditModal from '../../../components/SourceEditForm/SourceEditModal';
 import { routes, replaceRouteId } from '../../../Routes';
 import { applicationTypesData } from '../../applicationTypesData';
