+++ src/test/utilities/store.spec.js
@@ -1,9 +1,9 @@
-import { getProdStore, urlQueryMiddleware } from '../../Utilities/store';
+import { getProdStore, urlQueryMiddleware } from '../../utilities/store';
 import { defaultSourcesState } from '../../redux/sources/reducer';
 import { defaultUserState } from '../../redux/user/reducer';
-import * as queries from '../../Utilities/urlQuery';
+import * as queries from '../../utilities/urlQuery';
 import { ACTION_TYPES } from '../../redux/sources/actions-types';
-import { getDevStore } from '../../Utilities/getDevStore';
+import { getDevStore } from '../../utilities/getDevStore';
 
 describe('store creator', () => {
     const EXPECTED_DEFAULT_STATE = {
