+++ src/utilities.scss
@@ -1,0 +1,8 @@
+// Sass Imports
+
+@import 'utilities/_variables.scss';
+@import 'utilities/_colors.scss';
+@import 'utilities/_helpers.scss';
+@import 'utilities/_mixins.scss';
+
+@import '~@redhat-cloud-services/frontend-components/index.css';
